requirements.txt
yfinance
streamlit
pandas
python-dotenv
